# docs/swagger.yaml
openapi: 3.0.4
info:
  title: PPO Project.HttpServer
  version: v1
paths:
...
  '/api/v1/posts/{postId}':
    get:
      tags:
        - Post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/PostDto'
            application/json:
              schema:
                $ref: '#/components/schemas/PostDto'
            text/json:
              schema:
                $ref: '#/components/schemas/PostDto'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
    delete:
      tags:
        - Post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: No Content
          content:
            text/plain:
              schema:
                type: boolean
            application/json:
              schema:
                type: boolean
            text/json:
              schema:
                type: boolean
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
    patch:
      tags:
        - Post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePostDto'
          text/json:
            schema:
              $ref: '#/components/schemas/UpdatePostDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/UpdatePostDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/PostDto'
            application/json:
              schema:
                $ref: '#/components/schemas/PostDto'
            text/json:
              schema:
                $ref: '#/components/schemas/PostDto'
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
  /api/v1/posts:
    post:
      tags:
        - Post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostDto'
          text/json:
            schema:
              $ref: '#/components/schemas/CreatePostDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/CreatePostDto'
        required: true
      responses:
        '201':
          description: Created
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/PostDto'
            application/json:
              schema:
                $ref: '#/components/schemas/PostDto'
            text/json:
              schema:
                $ref: '#/components/schemas/PostDto'
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
  '/api/v1/companies/{companyId}/posts':
    get:
      tags:
        - Post
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PostDto'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PostDto'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PostDto'
        
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
  '/api/v1/employees/{employeeId}/postHistories/{postId}':
    get:
      tags:
        - PostHistory
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: postId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/PostHistoryDto'
            application/json:
              schema:
                $ref: '#/components/schemas/PostHistoryDto'
            text/json:
              schema:
                $ref: '#/components/schemas/PostHistoryDto'
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
    delete:
      tags:
        - PostHistory
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: employeeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: No Content
          content:
            text/plain:
              schema:
                type: boolean
            application/json:
              schema:
                type: boolean
            text/json:
              schema:
                type: boolean
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
    patch:
      tags:
        - PostHistory
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: employeeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePostHistoryDto'
          text/json:
            schema:
              $ref: '#/components/schemas/UpdatePostHistoryDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/UpdatePostHistoryDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/PostHistoryDto'
            application/json:
              schema:
                $ref: '#/components/schemas/PostHistoryDto'
            text/json:
              schema:
                $ref: '#/components/schemas/PostHistoryDto'
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
  /api/v1/postHistories:
    post:
      tags:
        - PostHistory
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostHistoryDto'
          text/json:
            schema:
              $ref: '#/components/schemas/CreatePostHistoryDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/CreatePostHistoryDto'
        required: true
      responses:
        '201':
          description: Created
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/PostHistoryDto'
            application/json:
              schema:
                $ref: '#/components/schemas/PostHistoryDto'
            text/json:
              schema:
                $ref: '#/components/schemas/PostHistoryDto'
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
  '/api/v1/employees/{employeeId}/postHistories':
    get:
      tags:
        - PostHistory
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: startDate
          in: query
          required: false
          schema:
            type: string
            format: date
            nullable: true
        - name: endDate
          in: query
          required: false
          schema:
            type: string
            format: date
            nullable: true
        - name: pageNumber
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
            default: 1
            nullable: false
        - name: pageSize
          in: query
          required: true
          schema:
            type: integer
            enum: [10, 20, 40]
            default: 10
            nullable: false
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PostDto'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PostDto'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PostDto'
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
...
components:
  schemas:
...
    CreatePostDto:
      required:
        - title
        - salary
        - companyId
      type: object
      properties:
        title:
          type: string
          nullable: false
        salary:
          type: number
          format: double
          nullable: false
        companyId:
          type: string
          format: uuid
          nullable: false
      additionalProperties: false
    CreatePostHistoryDto:
      required:
        - postId
        - employeeId
        - startDate
        - endDate
      type: object
      properties:
        postId:
          type: string
          format: uuid
          nullable: false
        employeeId:
          type: string
          format: uuid
          nullable: false
        startDate:
          type: string
          format: date
          nullable: false
        endDate:
          type: string
          format: date
          nullable: true
      additionalProperties: false
...
    ErrorDto:
      required:
        - errorType
        - message
      type: object
      properties:
        errorType:
          type: string
          nullable: false
        message:
          type: string
          nullable: true
      additionalProperties: false
...
    PostDto:
      required:
        - id
        - title
        - salary
        - companyId
        - isDeleted
      type: object
      properties:
        id:
          type: string
          format: uuid
          nullable: false
        title:
          type: string
          nullable: false
        salary:
          type: number
          format: double
          nullable: false
        companyId:
          type: string
          format: uuid
          nullable: false
        isDeleted:
          type: boolean
          nullable: false
      additionalProperties: false
    PostHistoryDto:
      required:
        - postId
        - employeeId
        - startDate
        - endDate
      type: object
      properties:
        postId:
          type: string
          format: uuid
          nullable: false
        employeeId:
          type: string
          format: uuid
          nullable: false
        startDate:
          type: string
          format: date
          nullable: false
        endDate:
          type: string
          format: date
          nullable: true
      additionalProperties: false
...
    UpdatePostDto:
      required:
        - salary
        - title
      type: object
      properties:
        title:
          type: string
          nullable: true
        salary:
          type: number
          format: double
          nullable: true
      additionalProperties: false
    UpdatePostHistoryDto:
      required:
        - startDate
        - endDate
      type: object
      properties:
        startDate:
          type: string
          format: date
          nullable: true
        endDate:
          type: string
          format: date
          nullable: true
      additionalProperties: false
...
  securitySchemes:
    Bearer:
      type: apiKey
      name: Authorization
      in: header
security:
  - Bearer: [ ]